Mejoras en scripts

+++ App/View/seleccionReglasDeNegocio.py
@@ -210,22 +210,34 @@ class Ui_Dialog(object):
         self.manageRulesButton.clicked.connect(self.open_manage_rules_dialog)
         
         selected_rules = self.main_window.controller.selected_rules
+        print(f"selected_rules: {selected_rules}")
+        existing_rules = set()
+
+        # Evitar reglas duplicadas en la tabla de opciones
+        for row in range(self.summaryOfOptionsTable.rowCount()):
+            rule_name = self.summaryOfOptionsTable.item(row, 0).text()
+            column_name = self.summaryOfOptionsTable.item(row, 1).text()
+            existing_rules.add((rule_name, column_name))
+
+        # Solo agregar reglas que no estén ya presentes en la tabla
         if selected_rules:
             for rule in selected_rules:
-                row_position = self.summaryOfOptionsTable.rowCount()
-                self.summaryOfOptionsTable.insertRow(row_position)
-                self.summaryOfOptionsTable.setItem(row_position, 0, QtWidgets.QTableWidgetItem(rule[0]))
-                self.summaryOfOptionsTable.setItem(row_position, 1, QtWidgets.QTableWidgetItem(rule[1]))
-                if isinstance(rule[2], dict):
-                    description = str(rule[2])  
-                else:
-                    description = rule[2]
-                    
-                self.summaryOfOptionsTable.setItem(row_position, 2, QtWidgets.QTableWidgetItem(description))  # Columna 2: Descripción de la regla
-                delete_button = QtWidgets.QPushButton("Eliminar")
-                delete_button.setStyleSheet("background-color: red; color: white;")
-                delete_button.clicked.connect(lambda checked, row=row_position: self.rule_selection_controller.remove_rule(self.summaryOfOptionsTable, row))
-                self.summaryOfOptionsTable.setCellWidget(row_position, 3, delete_button)
+                if (rule[0], rule[1]) not in existing_rules:  # Si la regla no está en la tabla
+                    row_position = self.summaryOfOptionsTable.rowCount()
+                    self.summaryOfOptionsTable.insertRow(row_position)
+                    self.summaryOfOptionsTable.setItem(row_position, 0, QtWidgets.QTableWidgetItem(rule[0]))
+                    self.summaryOfOptionsTable.setItem(row_position, 1, QtWidgets.QTableWidgetItem(rule[1]))
+                    if isinstance(rule[2], dict):
+                        description = str(rule[2])  # Convertir el dict a una cadena legible
+                    else:
+                        description = rule[2]  # Asumimos que es una cadena
+
+                    self.summaryOfOptionsTable.setItem(row_position, 2, QtWidgets.QTableWidgetItem(description))  # Columna 2: Descripción de la regla
+                    delete_button = QtWidgets.QPushButton("Eliminar")
+                    delete_button.setStyleSheet("background-color: red; color: white;")
+                    delete_button.setFont(QtGui.QFont("Segoe UI", 10, QtGui.QFont.Weight.Bold))
+                    delete_button.clicked.connect(lambda checked, row=row_position: self.controller.remove_rule(self.summaryOfOptionsTable, row))
+                    self.summaryOfOptionsTable.setCellWidget(row_position, 3, delete_button)
 
         self.retranslateUi(Dialog)
         QtCore.QMetaObject.connectSlotsByName(Dialog)
